cv/app/requirements.txt
djangorestframework-simplejwt
django-cors-headers
rest_framework_simplejwt